fix height swiper and rename components
