+++ src/app/page.jsx
@@ -1,13 +1,13 @@
 import Navbar from 'components/Navbar'
 import IntroSection from 'components/IntroSection'
 import ServicesSection from 'components/ServicesSection'
-import PortfilioSection from 'components/PortfilioSection'
+import PortfolioSection from 'components/PortfolioSection'
 import ContactUsSection from 'components/ContactUsSection'
 import AboutUsSection from 'components/AboutUsSection'
 import Footer from 'components/Footer'
-import styles from 'app/Home.module.scss'
 import Switcher from 'components/Switcher'
 import ReviewsSection from 'components/ReviewsSection'
+import styles from 'app/Home.module.scss'
 
 export default function Home() {
   return (
@@ -20,7 +20,7 @@ export default function Home() {
         <IntroSection />
         <AboutUsSection />
         <ServicesSection />
-        <PortfilioSection />
+        <PortfolioSection />
         <ReviewsSection />
         <ContactUsSection />
       </main>
