--- src/components/PortfilioSection/index.jsx
@@ -1,11 +1,0 @@
-import PortfolioBlock from 'components/PortfilioSection/PortfolioBlock'
-import WhyChooseBlock from 'components/PortfilioSection/WhyChooseBlock'
-
-export default function PortfilioSection() {
-  return (
-    <section id="portfolio">
-      <PortfolioBlock />
-      <WhyChooseBlock />
-    </section>
-  )
-}
