+++ src/components/PortfolioSection/index.jsx
@@ -1,0 +1,11 @@
+import PortfolioBlock from 'components/PortfolioSection/PortfolioBlock'
+import WhyChooseBlock from 'components/PortfolioSection/WhyChooseBlock'
+
+export default function PortfolioSection() {
+  return (
+    <section id="portfolio">
+      <PortfolioBlock />
+      <WhyChooseBlock />
+    </section>
+  )
+}
